requirements.txt
alembic==1.3.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
cffi==1.13.2
Click==7.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
plotly==4.5.0
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
retrying==1.3.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
Werkzeug==0.16.0
XlsxWriter==1.2.8
